requirements.txt
amqp==2.6.1
asgiref==3.4.1
billiard==3.6.4.0
celery==4.4.7
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.8
decorator==5.0.9
Django==3.2.6
djangorestframework==3.12.4
future==0.18.2
idna==3.2
kombu==4.6.11
Pillow==8.3.1
psycopg2-binary==2.8.6
pycparser==2.20
python-decouple==3.4
python-memcached==1.59
python-telegram-bot==12.5.1
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
tornado==6.1
urllib3==1.26.6
vine==1.3.0
whitenoise==5.3.0
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
coreapi==2.3.3
django-filter
